Update 2
-Cadastro salva
-Gosto quase salva
-BD.sql no projeto

+++ src/java/br/com/interestMe/service/ICategoriaService.java
@@ -2,6 +2,6 @@ package br.com.interestMe.service;
 
 import br.com.interestMe.entidade.Categoria;
 
-public interface ICategoriaService extends IGenericService<Categoria, Integer> {
+public interface ICategoriaService extends IGenericService<Categoria, Short> {
     
 }
