+++ src/java/br/com/interestMe/service/IGostoService.java
@@ -1,8 +1,7 @@
 package br.com.interestMe.service;
 
 import br.com.interestMe.entidade.Gosto;
-import java.util.List;
 
-public interface IGostoService extends IGenericService<Gosto, Integer>{
+public interface IGostoService extends IGenericService<Gosto, Short>{
     
 }
